requirements.txt
asgiref==3.3.4
cycler==0.10.0
Django==3.2.1
djangorestframework==3.12.4
kiwisolver==1.3.1
matplotlib==3.4.1
mysqlclient==2.0.3
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
